Fixed - XNATX-469, Added remote_project_id column in remoteMapper table to allow for switch in destination project. Fixed resource URI string.

+++ src/main/java/org/nrg/xsync/manager/SynchronizationManager.java
@@ -123,7 +123,8 @@ public class SynchronizationManager {
 	                  if (tmpF.isDirectory()) {
 	                      rmdir(tmpF);
 	                  }
-	                  tmpF.delete();
+	                  if (!tmpF.getAbsolutePath().endsWith("_sync.html"))
+	                	  tmpF.delete();
 	              }
 	          }
 	      }
